Smoke Detector
Our project is based on the detection of smoke. By using the IR sensor we will detect the smoke and with the help of alarm we can identify the smoke in the industries. If there is any presence of smoke in the industries the sensor will find the and start to make an alert. After the alarm sound we can identify the smoke proned areas and safeguard ourselves. With the help of this system we can safeguard the labours working in the industries. In this project we use the flame sensor is a type of sensor that can detect and respond to the presence of a flame. These detectors have the ability to identify smokeless liquid and smoke that can create open fire. For example, flame detectors are widely used, as a flame detector can detect heat, smoke, and fire. Application of Smoke alarms detect fires by sensing small particles in the air using a couple of different kinds of technologies. Once they detect those particles above a certain level, they signal the alarm to sound so that you and your family can get to safety. Arduino is the main board, which is an open-source microcontroller-based kit. The Arduino system offers a set of analog and digital pins that can be interfaced with many other boards and circuits which have different functions in a design. Figure 1 shows the Arduino Uno board. A sensor which is most sensitive to a normal light is known as a flame sensor. That’s why this sensor module is used in flame alarms. This sensor detects flame otherwise wavelength within the range of 760 nm – 1100 nm from the light source. This sensor can be easily damaged to high temperature. GSM module is a device that allows electronic devices to communicate with each other over the GSM network. GSM is a standard for digital cellular communications, which means that it provides a platform for mobile devices to communicate with each other wirelessly.